Updated Business.Web to HTMLDecode LinkedIn page data.  Updated WebTrakr database with update date fields.

+++ IODataBlock/Business.Web/Models/LinkedInCompanyScrapeDto.cs
@@ -5,7 +5,7 @@ namespace Business.Web.Models
     public class LinkedInCompanyScrapeDto
     {
         public string LocationUrl { get; set; }
-        public double CompanyId { get; set; }
+        public long CompanyId { get; set; }
         public string CompanyName { get; set; }
         public string DomainName { get; set; }
         public string Specialties { get; set; }
